requirements.txt
# this is the requirements.txt file

python-dotenv 

# see: https://github.com/theskumar/python-dotenv
